requirements.txt
psycopg2
django
djangorestframework
uwsgi
